fix(work-item): allow decimal targets and notify assignees on update
- Change target validation from integer to numeric (min:0.01) to accept
  decimal(10,2) values serialized as strings from DB
- Fix SyncWorkItemAssignmentsAction::syncAll type hint int → float
- Import and dispatch WorkItemUpdatedNotification in update() for all
  current assignees (skips the lead performing the update)
- Add InputError components to lead edit form for target/target_unit
- Update step/min attributes to 0.01 to allow decimal input

+++ app/Actions/WorkItem/SyncWorkItemAssignmentsAction.php
@@ -20,7 +20,7 @@ class SyncWorkItemAssignmentsAction
         }
     }
 
-    private function syncAll(WorkItem $workItem, Project $project, int $target, string $targetUnit): void
+    private function syncAll(WorkItem $workItem, Project $project, float $target, string $targetUnit): void
     {
         $memberIds = $project->members()->pluck('employees.id');
